Implement shader reflection on uniforms and ubos

+++ Project/include/Engine/App.h
@@ -16,7 +16,7 @@ public:
 	static void Init(int32_t argc, char** argv);
 	static App& Get(const std::string& name = "", uint16_t width = WINDOW_WIDTH, uint16_t height = WINDOW_HEIGHT);
 
-	float DeltaTime() const { return deltaTime; }
+	double DeltaTime() const { return deltaTime; }
 
 	void Run();
 private:
@@ -34,8 +34,8 @@ private:
 	Window* window;
 	Renderer* renderer;
 
-	float lastFrame = 0.0f;
-	float deltaTime = 0.0f;
+	double lastFrame = 0.;
+	double deltaTime = 0.;
 
 	friend class AppAttorney;
 };
